feat: 통계 서비스 모듈 추가 및 기본 배치 작업 설정

+++ video-service/src/main/java/com/sparta/videoservice/dto/VideoPlayResponse.java
@@ -15,5 +15,5 @@ public class VideoPlayResponse {
     private Long id;
     private Long videoId;
     private Long userId;
-    private Integer playDuration;
+    private Long playDuration;
 }
